+++ NangPaGo-be/src/main/java/com/mars/NangPaGo/config/SecurityConfig.java
@@ -3,6 +3,7 @@ package com.mars.NangPaGo.config;
 import com.mars.NangPaGo.domain.user.auth.CustomLogoutFilter;
 import com.mars.NangPaGo.domain.user.auth.CustomSuccessHandler;
 import com.mars.NangPaGo.domain.user.repository.RefreshTokenRepository;
+import com.mars.NangPaGo.domain.user.service.CustomLogoutService;
 import com.mars.NangPaGo.domain.user.service.CustomOAuth2UserService;
 import com.mars.NangPaGo.domain.user.util.JwtFilter;
 import com.mars.NangPaGo.domain.user.util.JwtUtil;
@@ -26,6 +27,7 @@ public class SecurityConfig {
 
     private final JwtUtil jwtUtil;
     private final CustomOAuth2UserService customOAuth2UserService;
+    private final CustomLogoutService customLogoutService;
     private final CustomSuccessHandler customSuccessHandler;
     private final RefreshTokenRepository refreshTokenRepository;
 
@@ -48,8 +50,9 @@ public class SecurityConfig {
             .authorizeHttpRequests(auth -> auth
                 .requestMatchers("/", "/login", "/oauth2/**", "/auth/reissue", "/auth/status").permitAll()
                 .anyRequest().authenticated());
-        http.addFilterBefore(new CustomLogoutFilter(jwtUtil, refreshTokenRepository), LogoutFilter.class);
+        http.addFilterBefore(new CustomLogoutFilter(customLogoutService), LogoutFilter.class);
         http
+
             .sessionManagement(session ->
                 session.sessionCreationPolicy(SessionCreationPolicy.STATELESS));
